Moving JuMP objects into their own file
